update dependency versions
`loom` 0.4.1 has a dependency on an old version of generator which is
subject to this security advisory:

https://rustsec.org/advisories/RUSTSEC-2020-0151.html

This causes `cargo deny` to report vulnerabilities in projects that
use `work-queue`.

Upgrading dependencies to newer versions to eliminate the known
vulnerabilities and reduce mixed versions (eg., using both loom 0.4.1
via work-queue and and loom 0.6.1 for newer projcets).